+++ src/DTO/Entity/Post/PostModerationDTO.php
@@ -1,19 +1,20 @@
 <?php
-namespace Blog\DTO\Entitie\Post;
+
+namespace Blog\DTO\Entity\Post;
 
 use Blog\Entity\Post;
 
-class PostModerationDTO{
-   public array $comments ;
-    public string $author ;
+class PostModerationDTO
+{
+    public array $comments;
+    public string $author;
     public string $title;
     public string $date;
     public function __construct(Post $post, array $comments)
     {
-        $this->author =$post->getUser()->getFullName();
+        $this->author = $post->getUser()->getFullName();
         $this->title = $post->getTitle();
         $this->date = \date_format($post->getDate(), "Y-m-d h:i:s");
         $this->comments = $comments;
     }
-    
 }
