+++ src/DTO/Form/Comment/CommentCreationDTO.php
@@ -1,9 +1,11 @@
 <?php
+
 namespace Blog\DTO\Form\Comment;
 
 use Blog\DTO\AbstractDTO;
 
-class CommentCreationDTO{
+class CommentCreationDTO
+{
     public string $content;
     public int $postId;
 }
