+++ src/DTO/Form/Comment/CommentModerationListDTO.php
@@ -1,11 +1,14 @@
 <?php
+
 declare(strict_types=1);
+
 namespace Blog\DTO\Form\Comment;
 
 use Blog\DTO\AbstractDTO;
 use Blog\Enum\CommentStatus;
 
-class CommentModerationListDTO {
+class CommentModerationListDTO
+{
     /**
      * @var array<int(commentid)>
      */
