+++ src/DTO/Form/Mail/MailDTO.php
@@ -1,6 +1,9 @@
 <?php
+
 namespace Blog\DTO\Form\Mail;
-class MailDTO{
+
+class MailDTO
+{
     public string $mail;
     public string $name;
     public string $message;
