+++ src/DTO/Form/Post/PostCreationDTO.php
@@ -1,6 +1,9 @@
 <?php
+
 namespace Blog\DTO\Form\Post;
-class PostCreationDTO{
+
+class PostCreationDTO
+{
     public string $title;
     public string $excerpt;
     public string $content;
