+++ src/DTO/Form/User/LoginDTO.php
@@ -1,12 +1,12 @@
 <?php
+
 namespace Blog\DTO\Form\User;
 
 use Blog\Enum\FieldType;
-use Blog\DTO\AbstractDTO;
 
-class LoginDTO{
+class LoginDTO
+{
     public string $login;
     public string $password;
     public FieldType $logintype;
-
 }
