+++ src/DTO/Form/User/RegisterDTO.php
@@ -1,13 +1,15 @@
 <?php
-namespace Blog\DTO\Form\User;
-use Blog\Enum\RoleEnum;
-use Blog\DTO\AbstractDTO;
 
-class RegisterDTO{
-        public string $login;
-        public string $password;
-        public string $mail;
-        public RoleEnum $role;
-        public string $firstName = "";
-        public string $lastName = "";
+namespace Blog\DTO\Form\User;
+
+use Blog\Enum\RoleEnum;
+
+class RegisterDTO
+{
+    public string $login;
+    public string $password;
+    public string $mail;
+    public RoleEnum $role;
+    public string $firstName = "";
+    public string $lastName = "";
 }
