--- src/DTO/Post/ListPostDTO.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Blog\DTO\Post;
-
-class ListPostDTO
-{
-    public array $listPost;
-}
