--- src/Service/Interface/DisplayerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Blog\Service\Interface;
-
-use Blog\DTO\AbstractDTO;
-
-interface DisplayerInterface
-{
-    public function display(int $id): object;
-}
