--- src/Service/Interface/GetterInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Blog\Service\Interface;
-
-interface GetterInterface
-{
-    public function getAll(): array;
-    public function getBy(array $params): object;
-}
