--- src/Service/Interface/LoggerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Blog\Service\Interface;
-
-use Blog\DTO\User\LoginDTO;
-
-interface LoggerInterface
-{
-    public function log(LoginDTO $objecttoCreate): int;
-}
