+++ src/Service/UserService.php
@@ -4,25 +4,19 @@ declare(strict_types=1);
 
 namespace Blog\Service;
 
-use Blog\DTO\Entitie\User\UserDTO;
 use Blog\Entity\User;
+use Blog\Enum\RoleEnum;
 use Doctrine\ORM\EntityManager;
 use Blog\DTO\Form\User\LoginDTO;
 use Blog\Exception\FormException;
-use Blog\Service\Interface\Logger;
+use Blog\DTO\Entity\User\UserDTO;
 use Doctrine\ORM\EntityRepository;
-use Blog\DTO\User\UserToDisplayDTO;
-use Blog\Enum\RoleEnum;
-use Blog\Service\Interface\Creater;
-use Blog\Service\Interface\Displayer;
+use Blog\DTO\Form\User\RegisterDTO;
 use Doctrine\Persistence\ObjectRepository;
-use Blog\Exception\UniqueKeyViolationException;
-use Blog\Service\Interface\DisplayerInterface;
-use Blog\Service\Interface\LoggerInterface;
-use Exception;
 
 
-class UserService 
+
+class UserService
 {
     private static ?UserService $_userService = null;
     private User $user;
@@ -84,6 +78,6 @@ class UserService
     }
     public function getUser(int $id): User
     {
-        return $this->entityManager->find(User::class, $id) ?? throw new Exception("no User with id" . (string)$id);
+        return $this->entityManager->find(User::class, $id) ?? throw new \Exception("no User with id" . (string)$id);
     }
 }
