+++ template/admin/createPost.html.twig
@@ -5,7 +5,7 @@
 
 	<div class="container">
         {% include "composant/error.html.twig" %}
-		{% if message %}
+		{% if post %}
 			<h2>{{post.title}} créer!</h2>
 			<p>
 				voulez vous créer un nouveau post</p>
